+++ adapters/Chartboost/ChartboostAdapter/GADMChartboostError.h
@@ -19,18 +19,11 @@
 #endif
 #import <Foundation/Foundation.h>
 
-/// Returns an NSError with NSLocalizedDescriptionKey and NSLocalizedFailureReasonErrorKey values
-/// set to |description|.
-NSError *GADChartboostErrorWithDescription(NSString *description);
-
-/// Returns an NSError with description acquired from the CBLoadError.
-NSError *adRequestErrorTypeForCBLoadError(CBLoadError error);
-
 /// Returns an NSError with description acquired from the CHBCacheError.
-NSError *NSErrorForCHBCacheError(CHBCacheError *error);
+NSError *GADMChartboostErrorForCHBCacheError(CHBCacheError *error);
 
 /// Returns an NSError with description acquired from the CHBShowError.
-NSError *NSErrorForCHBShowError(CHBShowError *error);
+NSError *GADMChartboostErrorForCHBShowError(CHBShowError *error);
 
 /// Returns an NSError with description acquired from the CHBClickError.
-NSError *NSErrorForCHBClickError(CHBClickError *error);
+NSError *GADMChartboostErrorForCHBClickError(CHBClickError *error);