+++ adapters/IronSource/IronSourceAdapter/GADMAdapterIronSourceConstants.h
@@ -14,7 +14,7 @@
 
 // IronSource internal reporting const.
 static NSString *const kGADMAdapterIronSourceMediationName = @"AdMob";
-static NSString *const kGADMAdapterIronSourceAdapterVersion = @"7.0.3.0";
+static NSString *const kGADMAdapterIronSourceAdapterVersion = @"7.1.0.0";
 static NSString *const kGADMIronSourceDefaultInstanceId = @"0";
 static NSString *const kGADMAdapterIronSourceInternalVersion = @"310";
 
@@ -23,8 +23,11 @@ static NSString *const kGADMAdapterIronSourceAppKey = @"appKey";
 static NSString *const kGADMAdapterIronSourceIsTestEnabled = @"isTestEnabled";
 static NSString *const kGADMAdapterIronSourceInstanceId = @"instanceId";
 
-// IronSource instance states
+// IronSource instance states.
 typedef NSString *InstanceState NS_STRING_ENUM;
 static InstanceState const kInstanceStateLocked = @"LOCKED";
 static InstanceState const kInstanceStateStart = @"START";
 static InstanceState const kInstanceStateCanLoad = @"CANLOAD";
+
+// IronSource mediation adapter error domain.
+static NSString *const kGADMAdapterIronSourceErrorDomain = @"com.google.mediation.IronSource";