+++ adapters/IronSource/IronSourceAdapter/GADMAdapterIronSourceUtils.m
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 #import "GADMAdapterIronSourceUtils.h"
+#import "GADMAdapterIronSourceConstants.h"
 
 void GADMAdapterIronSourceMutableSetAddObject(NSMutableSet *_Nullable set,
                                               NSObject *_Nonnull object) {
@@ -28,6 +29,16 @@ void GADMAdapterIronSourceMapTableSetObjectForKey(NSMapTable *_Nullable mapTable
   }
 }
 
+NSError *_Nonnull GADMAdapterIronSourceErrorWithCodeAndDescription(
+    GADMAdapterIronSourceErrorCode code, NSString *_Nonnull description) {
+  NSDictionary *userInfo =
+      @{NSLocalizedDescriptionKey : description, NSLocalizedFailureReasonErrorKey : description};
+  NSError *error = [NSError errorWithDomain:kGADMAdapterIronSourceErrorDomain
+                                       code:code
+                                   userInfo:userInfo];
+  return error;
+}
+
 @implementation GADMAdapterIronSourceUtils
 
 #pragma mark Utils Methods
@@ -37,18 +48,6 @@ void GADMAdapterIronSourceMapTableSetObjectForKey(NSMapTable *_Nullable mapTable
          ([value respondsToSelector:@selector(length)] && [(NSString *)value length] == 0) ||
          ([value respondsToSelector:@selector(length)] && [(NSData *)value length] == 0) ||
          ([value respondsToSelector:@selector(count)] && [(NSArray *)value count] == 0);
-}
-
-+ (NSError *)createErrorWith:(NSString *)description
-                   andReason:(NSString *)reason
-               andSuggestion:(NSString *)suggestion {
-  NSDictionary *userInfo = @{
-    NSLocalizedDescriptionKey : NSLocalizedString(description, nil),
-    NSLocalizedFailureReasonErrorKey : NSLocalizedString(reason, nil),
-    NSLocalizedRecoverySuggestionErrorKey : NSLocalizedString(suggestion, nil)
-  };
-
-  return [NSError errorWithDomain:NSStringFromClass([self class]) code:0 userInfo:userInfo];
 }
 
 + (void)onLog:(NSString *)log {
@@ -64,7 +63,7 @@ void GADMAdapterIronSourceMapTableSetObjectForKey(NSMapTable *_Nullable mapTable
     version = [version stringByReplacingOccurrencesOfString:@"." withString:@""];
 
   } @catch (NSException *exception) {
-    NSLog(@"Unable to parse AdMob SDK version");
+    NSLog(@"Unable to parse AdMob SDK version.");
     version = @"";
   }
   return version;