+++ adapters/Unity/UnityAdapter/GADMAdapterUnity.m
@@ -1,4 +1,4 @@
-// Copyright 2016 Google Inc.
+// Copyright 2020 Google LLC.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -13,16 +13,22 @@
 // limitations under the License.
 
 #import "GADMAdapterUnity.h"
-
 #import "GADMAdapterUnityBannerAd.h"
 #import "GADMAdapterUnityConstants.h"
-#import "GADMAdapterUnitySingleton.h"
 #import "GADMAdapterUnityUtils.h"
+#import "GADMUnityInterstitialAd.h"
 #import "GADMediationAdapterUnity.h"
+#import "GADUnityError.h"
 
-@interface GADMAdapterUnity () {
+@interface GADMAdapterUnity () <UnityAdsInitializationDelegate>
+@end
+
+@implementation GADMAdapterUnity {
   /// Connector from Google Mobile Ads SDK to receive ad configurations.
   __weak id<GADMAdNetworkConnector> _networkConnector;
+
+  /// Completion handler for initializing the Unity Ads SDK.
+  GADMediationAdapterSetUpCompletionBlock _initCompletionHandler;
 
   /// Game ID of Unity Ads network.
   NSString *_gameID;
@@ -33,16 +39,9 @@
   /// Unity Ads Banner wrapper
   GADMAdapterUnityBannerAd *_bannerAd;
 
-  /// UUID for Unity instrument analysis
-  NSString *_uuid;
-
-  /// MetaData for storing Unity instrument analysis
-  UADSMetaData *_metaData;
+  /// Unity Ads Interstitial Ad wrapper
+  GADMUnityInterstitialAd *_interstitialAd;
 }
-
-@end
-
-@implementation GADMAdapterUnity
 
 + (nonnull Class<GADMediationAdapter>)mainAdapterClass {
   return [GADMediationAdapterUnity class];
@@ -60,7 +59,31 @@
   if (_bannerAd != nil) {
     [_bannerAd stopBeingDelegate];
   }
-  [[GADMAdapterUnitySingleton sharedInstance] stopTrackingDelegate:self];
+}
+
+- (void)initializeWithGameID:(NSString *)gameID
+       withCompletionHandler:(GADMediationAdapterSetUpCompletionBlock)completionHandler {
+  if (![UnityAds isSupported]) {
+    NSString *message = [[NSString alloc] initWithFormat:@"%@ is not supported for this device.",
+                                                         NSStringFromClass([UnityAds class])];
+    NSError *error = GADMAdapterUnityErrorWithCodeAndDescription(
+        GADMAdapterUnityErrorAdInitializationFailure, message);
+    completionHandler(error);
+    return;
+  }
+
+  if ([UnityAds isInitialized]) {
+    NSLog(@"Unity Ads initialized successfully");
+    completionHandler(nil);
+    return;
+  }
+
+  // Configure metadata needed by Unity Ads SDK before initialization.
+  GADMAdapterUnityConfigureMediationService();
+
+  // Initializing Unity Ads with |gameID|.
+  _initCompletionHandler = completionHandler;
+  [UnityAds initialize:gameID testMode:NO enablePerPlacementLoad:YES initializationDelegate:self];
 }
 
 #pragma mark Interstitial Methods
@@ -69,66 +92,45 @@
   if (!connector) {
     return nil;
   }
-
   self = [super init];
   if (self) {
     _networkConnector = connector;
-
-    _uuid = [[NSUUID UUID] UUIDString];
-
-    _metaData = [[UADSMetaData alloc] init];
-
-    [_metaData setCategory:@"mediation_adapter"];
-    [_metaData set:_uuid value:@"create-adapter"];
-    [_metaData commit];
   }
   return self;
 }
 
 - (void)getInterstitial {
   id<GADMAdNetworkConnector> strongConnector = _networkConnector;
+  if (!strongConnector) return;
   _gameID = [[[strongConnector credentials] objectForKey:kGADMAdapterUnityGameID] copy];
-  _placementID = [[[strongConnector credentials] objectForKey:kGADMAdapterUnityPlacementID] copy];
-  if (!_gameID || !_placementID) {
-    NSError *error = GADMAdapterUnityErrorWithCodeAndDescription(
-        GADMAdapterUnityErrorInvalidServerParameters, @"Game ID and Placement ID cannot be nil.");
+  _interstitialAd = [[GADMUnityInterstitialAd alloc] initWithGADMAdNetworkConnector:strongConnector
+                                                                            adapter:self];
+  if (!_interstitialAd) {
+    NSString *description =
+        [NSString stringWithFormat:@"%@ initialization failed!",
+                                   NSStringFromClass([GADMUnityInterstitialAd class])];
+    NSError *error =
+        GADMAdapterUnityErrorWithCodeAndDescription(GADMAdapterUnityErrorAdObjectNil, description);
     [strongConnector adapter:self didFailAd:error];
     return;
   }
-
-  [_metaData setCategory:@"mediation_adapter"];
-  [_metaData set:_uuid value:@"load-interstitial"];
-  [_metaData set:_uuid value:_placementID];
-  [_metaData commit];
-
-  [[GADMAdapterUnitySingleton sharedInstance] requestInterstitialAdWithDelegate:self];
+  [_interstitialAd getInterstitial];
 }
 
 - (void)presentInterstitialFromRootViewController:(UIViewController *)rootViewController {
-  [_networkConnector adapterWillPresentInterstitial:self];
-
-  if (![UnityAds isReady:_placementID]) {
-    [_networkConnector adapterDidDismissInterstitial:self];
-    [_metaData setCategory:@"mediation_adapter"];
-    [_metaData set:_uuid value:@"fail-to-show-interstitial"];
-    [_metaData set:_uuid value:_placementID];
-    [_metaData commit];
-    return;
-  }
-
-  [_metaData setCategory:@"mediation_adapter"];
-  [_metaData set:_uuid value:@"show-interstitial"];
-  [_metaData set:_uuid value:_placementID];
-  [_metaData commit];
-  [[GADMAdapterUnitySingleton sharedInstance]
-      presentInterstitialAdForViewController:rootViewController
-                                    delegate:self];
+  [_interstitialAd presentInterstitialFromRootViewController:rootViewController];
 }
 
 #pragma mark Banner Methods
 
 - (void)getBannerWithSize:(GADAdSize)adSize {
   id<GADMAdNetworkConnector> strongConnector = _networkConnector;
+
+  if (!strongConnector) {
+    NSLog(@"Adapter Error: No GADMAdNetworkConnector found.");
+
+    return;
+  }
   GADAdSize supportedSize = [self supportedAdSizeFromRequestedSize:adSize];
   if (!IsGADAdSizeValid(supportedSize)) {
     NSString *errorMsg = [NSString
@@ -148,27 +150,19 @@
     [strongConnector adapter:self didFailAd:error];
     return;
   }
-
   _bannerAd = [[GADMAdapterUnityBannerAd alloc] initWithGADMAdNetworkConnector:strongConnector
                                                                        adapter:self];
-  [_bannerAd loadBannerWithSize:supportedSize];
-}
 
-#pragma mark GADMAdapterUnityDataProvider Methods
-
-- (NSString *)getGameID {
-  return _gameID;
-}
-
-- (NSString *)getPlacementID {
-  return _placementID;
-}
-
-- (void)didFailToLoadWithError:(nonnull NSError *)error {
-  id<GADMAdNetworkConnector> strongConnector = _networkConnector;
-  if (strongConnector != nil) {
+  if (!_bannerAd) {
+    NSString *description =
+        [NSString stringWithFormat:@"%@ initialization failed!",
+                                   NSStringFromClass([GADMAdapterUnityBannerAd class])];
+    NSError *error =
+        GADMAdapterUnityErrorWithCodeAndDescription(GADMAdapterUnityErrorAdObjectNil, description);
     [strongConnector adapter:self didFailAd:error];
+    return;
   }
+  [_bannerAd loadBannerWithSize:supportedSize];
 }
 
 /// Find closest supported ad size from a given ad size.
@@ -178,73 +172,22 @@
   return GADClosestValidSizeForAdSizes(gadAdSize, potentials);
 }
 
-#pragma mark - Unity Delegate Methods
+#pragma mark UnityAdsInitializationDelegate Methods
 
-- (void)unityAdsPlacementStateChanged:(NSString *)placementID
-                             oldState:(UnityAdsPlacementState)oldState
-                             newState:(UnityAdsPlacementState)newState {
-  if (![placementID isEqualToString:_placementID]) {
-    return;
-  }
-
-  id<GADMAdNetworkConnector> strongNetworkConnector = _networkConnector;
-  if (newState == kUnityAdsPlacementStateNoFill) {
-    if (strongNetworkConnector) {
-      NSString *errorMsg =
-          [NSString stringWithFormat:@"No ad available for the placement ID: %@", placementID];
-      NSError *error = GADMAdapterUnityErrorWithCodeAndDescription(
-          GADMAdapterUnityErrorPlacementStateNoFill, errorMsg);
-      [strongNetworkConnector adapter:self didFailAd:error];
-    }
-  }
-  if (newState == kUnityAdsPlacementStateDisabled) {
-    if (strongNetworkConnector) {
-      NSString *errorMsg =
-          [NSString stringWithFormat:@"This placement ID is currently disabled: %@", placementID];
-      NSError *error = GADMAdapterUnityErrorWithCodeAndDescription(
-          GADMAdapterUnityErrorPlacementStateDisabled, errorMsg);
-      [strongNetworkConnector adapter:self didFailAd:error];
-    }
+- (void)initializationComplete {
+  NSLog(@"Unity Ads initialized successfully");
+  if (_initCompletionHandler) {
+    _initCompletionHandler(nil);
   }
 }
 
-- (void)unityAdsDidFinish:(NSString *)placementID withFinishState:(UnityAdsFinishState)state {
-  if ([placementID isEqualToString:_placementID]) {
-    id<GADMAdNetworkConnector> strongNetworkConnector = _networkConnector;
-    if (strongNetworkConnector) {
-      [strongNetworkConnector adapterDidDismissInterstitial:self];
-    }
+- (void)initializationFailed:(UnityAdsInitializationError)error
+                 withMessage:(nonnull NSString *)message {
+  if (_initCompletionHandler) {
+    NSError *adapterError = GADMAdapterUnityErrorWithCodeAndDescription(
+        GADMAdapterUnityErrorAdInitializationFailure, message);
+    _initCompletionHandler(adapterError);
   }
-}
-
-- (void)unityAdsReady:(NSString *)placementID {
-  id<GADMAdNetworkConnector> strongNetworkConnector = _networkConnector;
-  if (strongNetworkConnector && [placementID isEqualToString:_placementID]) {
-    [strongNetworkConnector adapterDidReceiveInterstitial:self];
-  }
-}
-
-- (void)unityAdsDidClick:(NSString *)placementID {
-  id<GADMAdNetworkConnector> strongNetworkConnector = _networkConnector;
-  // The Unity Ads SDK doesn't provide an event for leaving the application, so the adapter assumes
-  // that a click event indicates the user is leaving the application for a browser or deeplink, and
-  // notifies the Google Mobile Ads SDK accordingly.
-  if (strongNetworkConnector && [placementID isEqualToString:_placementID]) {
-    [strongNetworkConnector adapterDidGetAdClick:self];
-    [strongNetworkConnector adapterWillLeaveApplication:self];
-  }
-}
-
-- (void)unityAdsDidError:(UnityAdsError)error withMessage:(NSString *)message {
-  id<GADMAdNetworkConnector> strongNetworkConnector = _networkConnector;
-  if (strongNetworkConnector) {
-    [strongNetworkConnector adapterWillDismissInterstitial:self];
-    [strongNetworkConnector adapterDidDismissInterstitial:self];
-  }
-}
-
-- (void)unityAdsDidStart:(nonnull NSString *)placementID {
-  // nothing to do
 }
 
 @end