+++ adapters/AppLovin/CHANGELOG.md
@@ -1,5 +1,35 @@
 ## AppLovin iOS Mediation Adapter Changelog
 
+#### Version 6.15.0.0
+- Verified compatibility with AppLovin SDK 6.15.0.
+
+Built and tested with:
+- Google Mobile Ads SDK version 7.69.0.
+- AppLovin SDK version 6.15.0.
+
+#### Version 6.14.11.0
+- Verified compatibility with AppLovin SDK 6.14.11.
+
+Built and tested with:
+- Google Mobile Ads SDK version 7.69.0.
+- AppLovin SDK version 6.14.11.
+
+#### Version 6.14.10.0
+- Verified compatibility with AppLovin SDK 6.14.10.
+- Now requires Google Mobile Ads SDK version 7.69.0 or higher.
+- Added the `GADMediationAdapterAppLovin.SDKSettings` property. Publishers may now configure AppLovin SDK settings through this API.
+
+Built and tested with:
+- Google Mobile Ads SDK version 7.69.0.
+- AppLovin SDK version 6.14.10.
+
+#### Version 6.14.9.0
+- Verified compatibility with AppLovin SDK 6.14.9.
+
+Built and tested with:
+- Google Mobile Ads SDK version 7.68.0.
+- AppLovin SDK version 6.14.9.
+
 #### Version 6.14.8.0
 - Verified compatibility with AppLovin SDK 6.14.8.